sample, root pkg tweak
