--- src/main/scala/magic_rdds/collect.scala
@@ -1,8 +1,0 @@
-package magic_rdds
-
-import org.hammerlab.magic.rdd.collect.CollectPartitions
-
-trait collect
-  extends CollectPartitions
-
-case object collect extends collect
