--- src/main/scala/magic_rdds/sort.scala
@@ -1,7 +1,0 @@
-package magic_rdds
-
-import org.hammerlab.magic.rdd
-
-trait sort extends rdd.sort
-object sort extends sort
-
