+++ src/main/scala/org/hammerlab/magic/rdd/collect.scala
@@ -1,10 +1,10 @@
-package org.hammerlab.magic.rdd.collect
+package org.hammerlab.magic.rdd
 
 import org.apache.spark.rdd.RDD
 
 import scala.reflect.ClassTag
 
-trait CollectPartitions {
+trait collect {
   implicit class CollectPartitionsOps[T: ClassTag](rdd: RDD[T]) extends Serializable {
     def collectParts: Array[Array[T]] =
       rdd
